Remove unnecessary compileClasspath on Environment
